+++ crates/tui/src/data_model/provider/local.rs
@@ -3,7 +3,8 @@ use serde::Deserialize;
 /// settings for local hardware as infra  
 #[derive(Debug, Default, Deserialize, Clone)]
 pub struct Settings {
-    pub docker_image: Option<String>,
-    pub commands: Option<Vec<String>>
+    pub docker_image: String,
+    pub mount_volume: String,
+    pub script: String
 }