+++ CHANGELOG.md
@@ -4,14 +4,14 @@ All notable changes will be documented in this file.
 ## [TODO]
 - [] job management
 - [] SSH accessible servers as a computation pod 
+- [] integrate quantum-expresso
 
 
 
 ## [Unreleased] v0.2.4
 ### Added
-- [] local as infra 
-- [] integrate quantum-expresso
-- [] add "version" at the seeting page
+- [x] local as infra 
+- [] add "version" at the setting page
 ### Changed
 - [x] fix root README.md
 - [] fix ubuntu 20.04 -> 24.04: [The Ubuntu 20.04 Actions runner image will begin deprecation on 2025-02-01 and will be fully unsupported by 2025-04-15](https://github.com/actions/runner-images/issues/11101)